Fixed groupId and bumped development version.
The Hudson update center is only scanning specific group-id:s
for hpi artifacts, com.sonyericsson.rebuild isn't one of them.
But it is scanning com.sonyericsson.hudson.plugins.*
which is what we should use as standard groupId from now on.
Since the Extension point is in this branch now I bumped the version to 1.1.